Add config and version subcommands
